move reports route under water systems to be more restful
