+++ spec/requests/reports_spec.rb
@@ -1,11 +1,11 @@
 require "rails_helper"
 
 RSpec.describe "Reports", type: :request do
-  describe "GET /reports/:pwsid" do
+  describe "GET /public_water_systems/:pwsid/report" do
     it "returns 200 with a report-body turbo frame" do
       pws = create(:public_water_system)
 
-      get "/reports/#{pws.pwsid}"
+      get "/public_water_systems/#{pws.pwsid}/report"
 
       expect(response).to have_http_status(:ok)
       expect(response.body).to include('id="report-body"')
@@ -14,7 +14,7 @@ RSpec.describe "Reports", type: :request do
     it "renders the system name and all section headings" do
       pws = create(:public_water_system, pws_name: "Clearwater Co")
 
-      get "/reports/#{pws.pwsid}"
+      get "/public_water_systems/#{pws.pwsid}/report"
 
       expect(response.body).to include("Clearwater Co")
       %w[Overview Demographics Violations Funding].each do |heading|
@@ -25,7 +25,7 @@ RSpec.describe "Reports", type: :request do
     it "handles nil associations gracefully" do
       pws = create(:public_water_system)
 
-      get "/reports/#{pws.pwsid}"
+      get "/public_water_systems/#{pws.pwsid}/report"
 
       expect(response).to have_http_status(:ok)
       expect(response.body).to include("Data not available")
@@ -36,14 +36,14 @@ RSpec.describe "Reports", type: :request do
       create(:demographic, pwsid: pws.pwsid, total_population: 75_000)
       create(:violations_summary, pwsid: pws.pwsid, health_violations_5yr: 3)
 
-      get "/reports/#{pws.pwsid}"
+      get "/public_water_systems/#{pws.pwsid}/report"
 
       expect(response.body).to include("75,000")
       expect(response.body).to include("3")
     end
 
     it "returns 404 when the system does not exist" do
-      get "/reports/DOESNOTEXIST"
+      get "/public_water_systems/ZZ0000000/report"
 
       expect(response).to have_http_status(:not_found)
     end
